--- k8s_zabbix_base/__init__.py
@@ -1,2 +1,0 @@
-
-
