+++ k8sobjects/daemonset.py
@@ -1,7 +1,6 @@
 import logging
 
-from pyzabbix import ZabbixMetric
-from .k8sobject import K8sObject, transform_value
+from .k8sobject import K8sObject
 
 logger = logging.getLogger(__name__)
