+++ k8sobjects/pod.py
@@ -2,7 +2,8 @@ import json
 import logging
 
 from pyzabbix import ZabbixMetric
-from .k8sobject import K8sObject, transform_value
+
+from .k8sobject import K8sObject
 
 logger = logging.getLogger(__name__)
