+++ mypy.ini
@@ -1,0 +1,4 @@
+[mypy]
+
+[mypy-sh,cachetools.*,pyzabbix.*,pause.*,coloredlogs.*,urllib3.*,requests,kubernetes]
+ignore_missing_imports = True
