+++ requirements.txt
@@ -1,6 +1,9 @@
-kubernetes==18.20.0
-cryptography==3.4.8
+kubernetes==19.15.0
+cryptography==35.0.0
 py-zabbix==1.1.7
-sentry-sdk==1.3.1
+sentry-sdk==1.5.0
 adal==1.2.7
-urllib3==1.26.6
+urllib3==1.26.7
+pytest==6.2.5
+mypy==0.910
+flake8==4.0.1
